requirements.txt
Flask==3.0.0
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0
dnspython==2.4.2
gunicorn==21.2.0
